Organize type definitions
