+++ src/types/index.d.ts
@@ -1,7 +1,9 @@
-export * from './activeBreakScene'
-export * from './activeRound'
-export * from './musicShown'
-export * from './nextRound'
-export * from './nextRoundStartTime'
-export * from './scoreboardData'
-export * from './swapColorsInternally'
+export * from './replicants/activeBreakScene'
+export * from './replicants/activeRound'
+export * from './replicants/musicShown'
+export * from './replicants/nextRound'
+export * from './replicants/nextRoundStartTime'
+export * from './replicants/scoreboardData'
+export * from './replicants/swapColorsInternally'
+export * from './replicants/obsData'
+export * from './replicants/gameAutomationData'
