+++ src/types/replicants/obsData.d.ts
@@ -1,0 +1,15 @@
+/* tslint:disable */
+/**
+ * This file was automatically generated by json-schema-to-typescript.
+ * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
+ * and run json-schema-to-typescript to regenerate this file.
+ */
+
+export interface ObsData {
+	enabled: boolean;
+	status: 'CONNECTED' | 'CONNECTING' | 'NOT_CONNECTED';
+	scenes?: string[];
+	currentScene?: string;
+	gameplayScene?: string;
+	intermissionScene?: string;
+}
